# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Chat:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      is_muted:
        type: boolean
      is_pinned:
        type: boolean
      last_message:
        type: string
      last_message_at:
        type: string
      last_message_id:
        type: integer
      last_name:
        type: string
      photo:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/domain.ChatType'
      unread_count:
        type: integer
      username:
        type: string
    type: object
  domain.ChatMessage:
    properties:
      chat_id:
        type: integer
      date:
        type: string
      forward_from:
        type: string
      from_id:
        type: integer
      from_name:
        type: string
      id:
        type: integer
      is_outgoing:
        type: boolean
      is_read:
        type: boolean
      media_type:
        description: photo, video, audio, document, sticker
        type: string
      media_url:
        type: string
      reply_to_id:
        type: integer
      text:
        type: string
    type: object
  domain.ChatType:
    enum:
    - private
    - group
    - supergroup
    - channel
    type: string
    x-enum-varnames:
    - ChatTypePrivate
    - ChatTypeGroup
    - ChatTypeSupergroup
    - ChatTypeChannel
  domain.ChatsResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/domain.Chat'
        type: array
      has_more:
        type: boolean
      total_count:
        type: integer
    type: object
  domain.Contact:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_blocked:
        type: boolean
      is_mutual:
        type: boolean
      last_name:
        type: string
      last_seen_at:
        type: string
      phone:
        type: string
      photo:
        type: string
      status:
        description: online, offline, recently, etc
        type: string
      username:
        type: string
    type: object
  domain.ContactsResponse:
    properties:
      contacts:
        items:
          $ref: '#/definitions/domain.Contact'
        type: array
      total_count:
        type: integer
    type: object
...
  domain.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  domain.HistoryResponse:
    properties:
      has_more:
        type: boolean
      messages:
        items:
          $ref: '#/definitions/domain.ChatMessage'
        type: array
      total_count:
        type: integer
    type: object
  domain.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  domain.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/domain.UserInfo'
    type: object
...
  domain.ResolveRequest:
    properties:
      phone:
        example: "+573001234567"
        type: string
      username:
        example: '@durov'
        type: string
    type: object
  domain.ResolvedPeer:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_bot:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/domain.ChatType'
      username:
        type: string
    type: object
  domain.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
...
  domain.UserInfo:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      username:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.MetaResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorResponse'
      meta:
        $ref: '#/definitions/handler.MetaResponse'
      success:
        type: boolean
    type: object
host: localhost:7789
info:
  contact: {}
  description: API REST para gestionar múltiples sesiones de Telegram via MTProto
  title: Telegram API
  version: 0.1.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica usuario y retorna tokens
      parameters:
      - description: Credenciales
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoca el refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Cerrar sesión
      tags:
      - Auth
  /auth/me:
    get:
      description: Retorna información del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Info usuario actual
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Genera nuevos tokens usando refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Renovar tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Crea una nueva cuenta de usuario
      parameters:
      - description: Datos del usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Registrar usuario
      tags:
      - Auth
...
  /sessions/{id}/chats:
    get:
      description: Obtiene la lista de chats/diálogos de la sesión
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Límite de resultados (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset para paginación
        in: query
        name: offset
        type: integer
      - description: Incluir chats archivados
        in: query
        name: archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Listar chats
      tags:
      - Chats
  /sessions/{id}/chats/{chatId}:
    get:
      description: Obtiene información detallada de un chat específico
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Chat'
              type: object
      security:
      - BearerAuth: []
      summary: Obtener información de chat
      tags:
      - Chats
  /sessions/{id}/chats/{chatId}/history:
    get:
      description: Obtiene el historial de mensajes de un chat
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - description: Límite de mensajes (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: ID del mensaje desde donde empezar
        in: query
        name: offset_id
        type: integer
      - description: Timestamp unix desde donde empezar
        in: query
        name: offset_date
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.HistoryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Obtener historial de mensajes
      tags:
      - Chats
  /sessions/{id}/contacts:
    get:
      description: Obtiene la lista de contactos de Telegram
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ContactsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Listar contactos
      tags:
      - Contacts
...
  /sessions/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Resuelve un @username o número de teléfono a un peer de Telegram
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Username o teléfono
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ResolvedPeer'
              type: object
      security:
      - BearerAuth: []
      summary: Resolver username o teléfono
      tags:
      - Contacts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"